prettier

+++ src/packages/relational/tests/delete/delete.spec.ts
@@ -8,25 +8,36 @@ import { greaterThan } from '../../sql/function/greater-than';
 
 describe('delete', () => {
   it('should delete without conditions', () => {
-    expectedSql({
-      delete: table(User),
-    }, 'DELETE FROM "auth"."user"');
+    expectedSql(
+      {
+        delete: table(User),
+      },
+      'DELETE FROM "auth"."user"',
+    );
   });
 
   it('should delete with equal condition', () => {
-    expectedSql({
-      delete: table(User),
-      where: equal(field(User, 'id'), '1')
-    }, 'DELETE FROM "auth"."user" WHERE "auth"."user"."id" = $1', ['1']);
+    expectedSql(
+      {
+        delete: table(User),
+        where: equal(field(User, 'id'), '1'),
+      },
+      'DELETE FROM "auth"."user" WHERE "auth"."user"."id" = $1',
+      ['1'],
+    );
   });
 
   it('should delete with equal condition', () => {
-    expectedSql({
-      delete: table(User),
-      where: or(
-        equal(field(User, 'id'), '1'),
-        greaterThan(field(User, 'loginCount'), 1)
-      )
-    }, 'DELETE FROM "auth"."user" WHERE ("auth"."user"."id" = $1 OR "auth"."user"."loginCount" > $2)', ['1', 1]);
+    expectedSql(
+      {
+        delete: table(User),
+        where: or(
+          equal(field(User, 'id'), '1'),
+          greaterThan(field(User, 'loginCount'), 1),
+        ),
+      },
+      'DELETE FROM "auth"."user" WHERE ("auth"."user"."id" = $1 OR "auth"."user"."loginCount" > $2)',
+      ['1', 1],
+    );
   });
 });
